+++ spacy/tests/conftest.py
@@ -1,6 +1,4 @@
 import pytest
-
-from spacy.tokens import Doc
 from spacy.util import get_lang_class
 
 
@@ -470,25 +468,3 @@ def zh_tokenizer_pkuseg():
 @pytest.fixture(scope="session")
 def hy_tokenizer():
     return get_lang_class("hy")().tokenizer
-
-
-@pytest.fixture()
-def doc(en_vocab):
-    words = ["c", "d", "e"]
-    pos = ["VERB", "NOUN", "NOUN"]
-    tags = ["VBP", "NN", "NN"]
-    heads = [0, 0, 0]
-    deps = ["ROOT", "dobj", "dobj"]
-    ents = ["O", "B-ORG", "O"]
-    morphs = ["Feat1=A", "Feat1=B", "Feat1=A|Feat2=D"]
-
-    return Doc(
-        en_vocab,
-        words=words,
-        pos=pos,
-        tags=tags,
-        heads=heads,
-        deps=deps,
-        ents=ents,
-        morphs=morphs,
-    )
